Refactor _LockedLobby into JsonLock
